# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "uiscloud_weknora 지식베이스 관리 시스템 API 문서",
        "title": "uiscloud_weknora API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "uiscloud_weknora Github",
            "url": "https://github.com/rockgis/uiscloud_weknora"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/mcp-services": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "현재 테넌트의 모든 MCP 서비스를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 목록 조회",
                "responses": {
                    "200": {
                        "description": "MCP 서비스 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "새 MCP 서비스 설정을 생성합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 생성",
                "parameters": [
                    {
                        "description": "MCP 서비스 설정",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "생성된 MCP 서비스",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/mcp-services/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "ID로 MCP 서비스 상세 정보를 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP 서비스 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "MCP 서비스 상세 정보",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "서비스가 존재하지 않습니다",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "MCP 서비스 설정을 업데이트합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 업데이트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP 서비스 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "업데이트 필드",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "업데이트된 MCP 서비스",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "지정된 MCP 서비스를 삭제합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP 서비스 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/mcp-services/{id}/resources": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "MCP 서비스가 제공하는 리소스 목록을 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 리소스 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP 서비스 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "리소스 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/mcp-services/{id}/test": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "MCP 서비스의 정상 연결 여부를 테스트합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 연결 테스트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP 서비스 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "테스트 결과",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 파라미터",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
        "/mcp-services/{id}/tools": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "MCP 서비스가 제공하는 도구 목록을 조회합니다",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP 서비스"
                ],
                "summary": "MCP 서비스 도구 목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP 서비스 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "도구 목록",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_Tencent_uiscloud_weknora_internal_errors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_errors.ErrorCode": {
            "type": "integer",
            "enum": [
                1000,
                1001,
                1002,
                1003,
                1004,
                1005,
                1006,
                1007,
                1008,
                1009,
                1010,
                2000,
                2001,
                2002,
                2003,
                2004,
                2100,
                2101,
                2102,
                2103
            ],
            "x-enum-varnames": [
                "ErrBadRequest",
                "ErrUnauthorized",
                "ErrForbidden",
                "ErrNotFound",
                "ErrMethodNotAllowed",
                "ErrConflict",
                "ErrTooManyRequests",
                "ErrInternalServer",
                "ErrServiceUnavailable",
                "ErrTimeout",
                "ErrValidation",
                "ErrTenantNotFound",
                "ErrTenantAlreadyExists",
                "ErrTenantInactive",
                "ErrTenantNameRequired",
                "ErrTenantInvalidStatus",
                "ErrAgentMissingThinkingModel",
                "ErrAgentMissingAllowedTools",
                "ErrAgentInvalidMaxIterations",
                "ErrAgentInvalidTemperature"
            ]
        },
...
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPAdvancedConfig": {
            "type": "object",
            "properties": {
                "retry_count": {
                    "description": "Number of retries, default: 3",
                    "type": "integer"
                },
                "retry_delay": {
                    "description": "Delay between retries in seconds, default: 1",
                    "type": "integer"
                },
                "timeout": {
                    "description": "Timeout in seconds, default: 30",
                    "type": "integer"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPAuthConfig": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "custom_headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPEnvVars": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPHeaders": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPService": {
            "type": "object",
            "properties": {
                "advanced_config": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPAdvancedConfig"
                },
                "auth_config": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPAuthConfig"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "env_vars": {
                    "description": "Environment variables for stdio",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPEnvVars"
                        }
                    ]
                },
                "headers": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPHeaders"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stdio_config": {
                    "description": "Required for stdio transport",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPStdioConfig"
                        }
                    ]
                },
                "tenant_id": {
                    "type": "integer"
                },
                "transport_type": {
                    "$ref": "#/definitions/github_com_Tencent_uiscloud_weknora_internal_types.MCPTransportType"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "description": "Optional: required for SSE/HTTP Streamable",
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPStdioConfig": {
            "type": "object",
            "properties": {
                "args": {
                    "description": "Command arguments array",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "command": {
                    "description": "Command: \"uvx\" or \"npx\"",
                    "type": "string"
                }
            }
        },
        "github_com_Tencent_uiscloud_weknora_internal_types.MCPTransportType": {
            "type": "string",
            "enum": [
                "sse",
                "http-streamable",
                "stdio"
            ],
            "x-enum-comments": {
                "MCPTransportHTTPStreamable": "HTTP Streamable",
                "MCPTransportSSE": "Server-Sent Events",
                "MCPTransportStdio": "Stdio (Standard Input/Output)"
            },
            "x-enum-descriptions": [
                "Server-Sent Events",
                "HTTP Streamable",
                "Stdio (Standard Input/Output)"
            ],
            "x-enum-varnames": [
                "MCPTransportSSE",
                "MCPTransportHTTPStreamable",
                "MCPTransportStdio"
            ]
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer {token} 형식의 JWT 토큰을 입력하세요",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}